ti-common: wipe all eeprom when updating kv3 serial
Userland is using the remainder of eeprom not reserved for the
bootloader as a permanent data store. However, some of the data being
stored is directly tied to the device serial number. As the serial
number is only ever set during manufacturing, wipe eeprom each time each
time it is changed.