ti: am335x: Fix algorithm that determines LCD init sequence
There are two different LCDs that require different initialization
sequences.  The basic steps that are used to determine which LCD is
there are as follows:
	1) Read the hardware revision by reading gpio_1_26 (bit 0),
	   gpio1_27 (bit 1), gpio1_28 (bit 2), ad gpio1_29 (bit 3),
	   invert the bits, and if the resulting value is 3 or 4,
	   then it is a version 1 LCD.  If the resulting value is 5,
	   then it is a version 2 LCD.  If it is some other value,
	   go to step 2.
	2) Read the LCD version by reading gpio3_0 (bit 0) and
	   gpio3_1 (bit) and if the resulting value is 3, then it is
	   a version 1 LCD.  If the resulting value is 2, then it is
	   a version 2 LCD.  If it is neither, then we don't know what
	   it is so default to a version 2 LCD.

In order for this to work, the pinmux needs to be set up to enable
GPIO for the gpmc_a10, gpmc_a11, gpmc_be1n, and gpmc_csn0 pins which
are used to access the hardware version.

Also, rename the GPIO_REVISION[01] macros to GPIO_LCD_REV[01] to
make their purpose clearer.

Signed-off-by: Mark Greer <mgreer@animalcreek.com>